requirements.txt
streamlit
groq
feedparser
python-dotenv
pandas
plotly
fpdf2